+++ comedores/forms/comedor_form.py
@@ -1,17 +1,10 @@
 from django import forms
 
 
-from comedores.models.comedor import (
-    Comedor,
-    Referente,
-    Intervencion,
-    ImagenComedor,
-    Nomina,
-)
+from comedores.models import Comedor, Referente, Intervencion, ImagenComedor, Nomina
 
 from configuraciones.models import Municipio, Provincia
 from configuraciones.models import Localidad
-from organizaciones.models import Organizacion
 
 
 class ReferenteForm(forms.ModelForm):
