+++ comedores/serializers/relevamiento_serializer.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
-from comedores.models.relevamiento import Relevamiento
-from comedores.models.relevamiento import (
+from comedores.models import (
+    Relevamiento,
     Territorial,
 )
 from comedores.services.relevamiento_service import RelevamientoService
@@ -17,7 +17,7 @@ class RelevamientoSerializer(serializers.ModelSerializer):
 
         if "territorial" in self.initial_data:
             territorial_data = self.initial_data["territorial"]
-            territorial, _ = Territorial.objects.get_or_create(
+            territorial, _created = Territorial.objects.get_or_create(
                 gestionar_uid=territorial_data["gestionar_uid"],
                 defaults={"nombre": territorial_data["nombre"]},
             )
