+++ comedores/services/relevamiento_service.py
@@ -1,29 +1,23 @@
 import os
-import json
-from django.shortcuts import get_object_or_404
 from django.utils import timezone
 import requests
 
 from django.db import models
-
-from comedores.models.relevamiento import Relevamiento
-from comedores.models.comedor import (
-    Comedor,
-)
-from comedores.models.relevamiento import (
+from comedores.models import (
     Anexo,
     CantidadColaboradores,
-    Colaboradores,
     Espacio,
     EspacioCocina,
     EspacioPrestacion,
+    Colaboradores,
     FrecuenciaLimpieza,
     FrecuenciaRecepcionRecursos,
     FuenteCompras,
     FuenteRecursos,
     FuncionamientoPrestacion,
     Prestacion,
-    Territorial,
+    Relevamiento,
+    TipoAccesoComedor,
     TipoAgua,
     TipoCombustible,
     TipoDesague,
@@ -35,7 +29,6 @@ from comedores.models.relevamiento import (
     TipoModalidadPrestacion,
     TipoRecurso,
     TipoTecnologia,
-    TipoAccesoComedor,
 )
 
 
@@ -43,48 +36,6 @@ class RelevamientoService:
     @staticmethod
     def convert_to_int(value):
         return int(value) if value != "" else None
-    
-    @staticmethod
-    def create_pendiente(request, comedor_id):
-        comedor = get_object_or_404(Comedor, id=comedor_id)
-        relevamiento = Relevamiento(comedor=comedor, estado="Pendiente")
-        territorial_data = request.POST.get("territorial")
-
-        if territorial_data:
-            territorial_data = json.loads(territorial_data)
-            gestionar_uid = territorial_data.get("gestionar_uid")
-            nombre = territorial_data.get("nombre")
-
-            territorial, _ = Territorial.objects.get_or_create(
-                gestionar_uid=gestionar_uid, defaults={"nombre": nombre}
-            )
-            relevamiento.territorial = territorial
-            relevamiento.estado = "Visita pendiente"
-
-        relevamiento.save()
-
-        return relevamiento
-
-    @staticmethod
-    def update_territorial(request):
-        relevamiento_id = request.POST.get("relevamiento_id")
-        relevamiento = Relevamiento.objects.get(id=relevamiento_id)
-        territorial_data = request.POST.get("territorial_editar")
-
-        if territorial_data:
-            territorial_data = json.loads(territorial_data)
-            gestionar_uid = territorial_data["gestionar_uid"]
-            territorial = Territorial.objects.get(gestionar_uid=gestionar_uid)
-
-            relevamiento.territorial = territorial
-            relevamiento.estado = "Visita pendiente"
-        else:
-            relevamiento.territorial = None
-            relevamiento.estado = "Pendiente"
-
-        relevamiento.save()
-
-        return relevamiento
 
     @staticmethod
     def populate_relevamiento(relevamiento_form, extra_forms, user_id):
@@ -1198,4 +1149,4 @@ class RelevamientoService:
                 relevamiento.docPDF = response["Rows"][0]["docPDF"]
                 relevamiento.save()
             except requests.exceptions.RequestException as e:
-                print(f"Error al sincronizar con GESTIONAR: {e}")
+                print(f"Error en la petición POST: {e}")
