+++ comedores/templates/comedor/comedor_form.html
@@ -50,10 +50,7 @@
                                 <div class="form-group row">
                                     <div class="col-3">{{ form.nombre|as_crispy_field }}</div>
                                     <div class="col-3">{{ form.tipocomedor|as_crispy_field }}</div>
-                                    <div class="col-3">{{ form.organizacion|as_crispy_field }}</div>
-                                    <div class="col-3">{{ form.comienzo|as_crispy_field }}</div>
-                                </div>
-                                <div class="form-group row">
+                                    <div class="col-2">{{ form.comienzo|as_crispy_field }}</div>
                                     <div class="col-3">{{ form.provincia|as_crispy_field }}</div>
                                     <div class="col-4">{{ form.municipio|as_crispy_field }}</div>
                                 </div>
