+++ config/settings.py
@@ -129,7 +129,6 @@ INSTALLED_APPS = [
     "configuraciones",
     "dashboard",
     "comedores",
-    "organizaciones",
     "provincias",
 ]
 
@@ -147,6 +146,7 @@ MIDDLEWARE = [
     "debug_toolbar.middleware.DebugToolbarMiddleware",
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
     "config.middlewares.xss_protection.XSSProtectionMiddleware",
+    "config.middlewares.no_special_characters.NoSpecialCharactersMiddleware",
     "corsheaders.middleware.CorsMiddleware",
 ]
