--- organizaciones/templates/organizacion_confirm_delete.html
@@ -1,37 +1,0 @@
-{% extends "includes/main.html" %};
-{% load static %}
-;
-{% block title %}Organizacion - {{ organizacion.nombre }}{% endblock %}
-;
-{% block breadcrumb %}
-    <ol class="breadcrumb float-sm-right">
-        <div class="breadcrumb-item">
-            <a href="{% url 'organizaciones' %}">Organizaciones</a>
-        </div>
-        <div class="breadcrumb-item">
-            <a href="{% url 'organizacion_detalle' organizacion.id %}">{{ organizacion.nombre }}</a>
-        </div>
-        <div class="breadcrumb-item">Eliminar</div>
-    </ol>
-{% endblock %}
-;
-{% block content %}
-    <div class="error-page mt-5 px-3">
-        <div class="error-content">
-            <h3>
-                <i class="fas fa-exclamation-triangle text-danger px-2"></i> Atención!
-            </h3>
-            <p>
-                ¿Estás seguro que deseas eliminar de manera permanente el organizacion: <span class="font-weight-bold"><a href="{% url 'organizacion_detalle' organizacion.id %}">{{ organizacion.nombre }}</a></span>?
-            </p>
-            <div class="text-left">
-                <form method="post">
-                    {% csrf_token %}
-                    <button type="submit" class="btn btn-danger btn-sm">Eliminar</button>
-                    <a href="{% url 'organizacion_detalle' organizacion.id %}"
-                       class="btn btn-secondary btn-sm">Cancelar</a>
-                </form>
-            </div>
-        </div>
-    </div>
-{% endblock %}
