--- organizaciones/templates/organizacion_form.html
@@ -1,64 +1,0 @@
-{% extends "includes/main.html" %}
-{% load static crispy_forms_tags %}
-;
-{% block title %}Organizaciones{% endblock %}
-{% block titulo-pagina %}Organizaciones{% endblock %}
-{% block breadcrumb %}
-    <ol class="breadcrumb float-sm-right px-3">
-        <li class="breadcrumb-item">
-            <a href="{% url 'organizaciones' %}">Organizaciones</a>
-        </li>
-        {% if request.resolver_match.url_name == "organizacion_editar" %}
-            <li class="breadcrumb-item">
-                <a href="{% url 'organizacion_detalle' organizacion.id %}">{{ organizacion.nombre }}</a>
-            </li>
-            <li class="breadcrumb-item active">Editar</li>
-        {% else %}
-            <li class="breadcrumb-item active">Agregar</li>
-        {% endif %}
-    </ol>
-{% endblock %}
-;
-{% block content %}
-    <div class="d-flex justify-content-between">
-        <div>
-            {% if organizacion.id %}
-                <a href="{% url 'organizacion_detalle' organizacion.id %}"
-                   class='btn btn-secondary'>Volver</a>
-            {% else %}
-                <a href="{% url 'organizaciones' %}" class='btn btn-secondary'>Volver</a>
-            {% endif %}
-        </div>
-        <div>
-            <a class='btn btn-secondary print mr-1 d-none d-sm-inline'>Imprimir</a>
-            <a href='#' onclick='window.history.back();' class='btn btn-secondary'>Cancelar</a>
-        </div>
-    </div>
-    <div class="row organizacion-form">
-        <div class="col-12">
-            <form method="post" class="w-100" enctype="multipart/form-data">
-                <div class="card card-primary card-outline my-3">
-                    <div class="bs-stepper">
-                        <div class="card-header mb-3">
-                            <h3 class="card-title">Agregar Organizacion</h3>
-                        </div>
-                        <div class="card-body">
-                            <div class="form-row justify-content-center">
-                                {% csrf_token %}
-                                <div class="form-group row">
-                                    <div class="col-3">{{ form.nombre|as_crispy_field }}</div>
-                                    <div class="col-3">{{ form.cuit|as_crispy_field }}</div>
-                                    <div class="col-3">{{ form.telefono|as_crispy_field }}</div>
-                                    <div class="col-3">{{ form.email|as_crispy_field }}</div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="d-flex justify-content-end">
-                    <button class="btn btn-success mx-1" type="submit">Guardar organizacion</button>
-                </div>
-            </form>
-        </div>
-    </div>
-{% endblock %}
