--- organizaciones/templates/organizacion_list.html
@@ -1,127 +1,0 @@
-{% extends "includes/main.html" %};
-{% load static %}
-;
-{% block title %}Organiaciones{% endblock %}
-;
-{% block breadcrumb %}
-    <ol class="breadcrumb float-sm-right">
-        <li class="breadcrumb-item">
-            <a href="{% url 'organizaciones' %}">Organiaciones</a>
-        </li>
-        <li class="breadcrumb-item active">Listar</li>
-    </ol>
-{% endblock %}
-;
-{% block content %}
-    <h2 class="text-center display-4 mb-4">Buscar Organiacion</h2>
-    <form method="GET">
-        <div class="row">
-            <div class="col-md-10 offset-md-1">
-                <div class="form-group">
-                    <div class="input-group input-group-lg">
-                        <input type="search"
-                               class="form-control form-control-lg"
-                               placeholder="Buscar por nombre, CUIT, telefono o email"
-                               title="Ingresar nombre, CUIT, telefono o email."
-                               name="busqueda" />
-                        <div class="input-group-append">
-                            <button type="submit" class="btn btn-lg btn-primary">
-                                <i class="fa fa-search"></i>
-                            </button>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </form>
-    <div class="row mt-3">
-        <div class="col-12 text-center">
-            <a href="{% url 'organizaciones' %}"
-               class="btn btn-secondary btn-lg d-block d-sm-inline mt-2 ">Resetear</a>
-            <a href="{% url 'organizacion_crear' %}"
-               class="btn btn-primary btn-lg d-block d-sm-inline mt-2 mr-2 ">Agregar</a>
-        </div>
-    </div>
-    <div class="row mt-5">
-        <div class="col-12">
-            <div class="card">
-                <div class="card-body">
-                    <table class="table table-bordered table-striped projects">
-                        <thead>
-                            <tr>
-                                <th>Nombre</th>
-                                <th>CUIT</th>
-                                <th>Telefono</th>
-                                <th>Email</th>
-                                <th style="width: 15%" class="notexport">Aciones</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {% for organizacion in organizaciones %}
-                                <tr>
-                                    <td>
-                                        <a href="{% url 'organizacion_detalle' organizacion.id %}"
-                                           class="font-weight-bold"
-                                           title="Ver detalles">{{ organizacion.nombre }}</a>
-                                    </td>
-                                    <td>{{ organizacion.cuit |default_if_none:"-" }}</td>
-                                    <td>{{ organizacion.telefono | default_if_none:"-" }}</td>
-                                    <td>{{ organizacion.email | default_if_none:"-" }}</td>
-                                    <td>
-                                        <div class="btn-group" role="group">
-                                            <a href="{% url 'organizacion_detalle' organizacion.id %}"
-                                               class="btn btn-primary btn-sm ver">Ver</a>
-                                            <a href="{% url 'organizacion_editar' organizacion.id %}"
-                                               class="btn btn-primary btn-sm editar">Editar</a>
-                                            <a href="{% url 'organizacion_eliminar' organizacion.id %}"
-                                               class="btn btn-danger btn-sm eliminar">Eliminar</a>
-                                        </div>
-                                    </td>
-                                </tr>
-                            {% endfor %}
-                        </tbody>
-                    </table>
-                    {% if is_paginated %}
-                        <div class="mt-2">
-                            <ul class="pagination">
-                                {% if page_obj.has_previous %}
-                                    <li class="page-item">
-                                        <a href="?{% if query %}busqueda={{ query }}&{% endif %}page={{ page_obj.previous_page_number }}"
-                                           class="page-link">Volver</a>
-                                    </li>
-                                {% else %}
-                                    <li class="page-item disabled">
-                                        <a href="#" class="page-link">Volver</a>
-                                    </li>
-                                {% endif %}
-                                {% for i in page_range %}
-                                    {% if i == '…' %}
-                                        <li class="page-item disabled">
-                                            <a href="#" class="page-link">…</a>
-                                        </li>
-                                    {% else %}
-                                        <li class="page-item {% if page_obj.number == i %}active{% endif %}">
-                                            <a href="?{% if query %}busqueda={{ query }}&{% endif %}page={{ i }}"
-                                               class="page-link">{{ i }}</a>
-                                        </li>
-                                    {% endif %}
-                                {% endfor %}
-                                {% if page_obj.has_next %}
-                                    <li class="page-item">
-                                        <a href="?{% if query %}busqueda={{ query }}&{% endif %}page={{ page_obj.next_page_number }}"
-                                           class="page-link">Continuar</a>
-                                    </li>
-                                {% else %}
-                                    <li class="page-item disabled">
-                                        <a href="#" class="page-link">Continuar</a>
-                                    </li>
-                                {% endif %}
-                            </ul>
-                        </div>
-                    {% endif %}
-                </div>
-            </div>
-        </div>
-    </div>
-{% endblock %}
-{% block customJS %}{% endblock %}
