--- organizaciones/views.py
@@ -1,64 +1,0 @@
-from django.db.models import Q
-from django.urls import reverse, reverse_lazy
-from django.views.generic import (
-    CreateView,
-    DeleteView,
-    DetailView,
-    ListView,
-    UpdateView,
-)
-
-from organizaciones.forms import OrganizacionForm
-from organizaciones.models import Organizacion
-
-
-class OrganizacionListView(ListView):
-    model = Organizacion
-    template_name = "organizacion_list.html"
-    context_object_name = "organizaciones"
-    paginate_by = 10
-
-    def get_queryset(self):
-        query = self.request.GET.get("busqueda")
-        queryset = Organizacion.objects.all()
-
-        if query:
-            queryset = queryset.filter(
-                Q(nombre__icontains=query)
-                | Q(cuit__icontains=query)
-                | Q(telefono__icontains=query)
-                | Q(email__icontains=query)
-            )
-
-        return queryset
-
-
-class OrganizacionCreateView(CreateView):
-    model = Organizacion
-    form_class = OrganizacionForm
-    template_name = "organizacion_form.html"
-
-    def get_success_url(self):
-        return reverse("organizacion_detalle", kwargs={"pk": self.object.pk})
-
-
-class OrganizacionUpdateView(UpdateView):
-    model = Organizacion
-    form_class = OrganizacionForm
-    template_name = "organizacion_form.html"
-
-    def get_success_url(self):
-        return reverse("organizacion_detalle", kwargs={"pk": self.object.pk})
-
-
-class OrganizacionDetailView(DetailView):
-    model = Organizacion
-    template_name = "organizacion_detail.html"
-    context_object_name = "organizacion"
-
-
-class OrganizacionDeleteView(DeleteView):
-    model = Organizacion
-    template_name = "organizacion_confirm_delete.html"
-    context_object_name = "organizacion"
-    success_url = reverse_lazy("organizaciones")
